% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anocva.R
\name{checkClusteringFunction}
\alias{checkClusteringFunction}
\title{Verifies if a clustering function is set. If not,
uses the spectral clustering as default clustering method.}
\usage{
checkClusteringFunction(clusteringFunction)
}
\arguments{
\item{clusteringFunction}{The actual clustering function. NULL if its unknown.}
}
\value{
A clustering function
}
\description{
Verifies if a clustering function is set. If not,
uses the spectral clustering as default clustering method.
}
\keyword{internal}
