% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nClust.R
\name{optimalSlope}
\alias{optimalSlope}
\title{Estimates the optimal number of clusters using the Slope criterion when silhouette statistics already exists.
The optimal number of clusters will be verified in the range {2,..., maxClust}.}
\usage{
optimalSlope(silStats, p = 1, maxClust)
}
\arguments{
\item{silStats}{Silhouette statistics.}

\item{p}{Slope adjust parameter.}

\item{maxClust}{The max number of clusters to be tried.}
}
\value{
The optimal number of clusters.
}
\description{
Estimates the optimal number of clusters using the Slope criterion when silhouette statistics already exists.
The optimal number of clusters will be verified in the range {2,..., maxClust}.
}
\references{
Fujita A, Takahashi DY, Patriota AG (2014b) A non-parametric method to
estimate the number of clusters. Computational Statistics & Data Analysis 73:27–39
}
\keyword{internal}
